Lecture 312. Delete Customer - Creating Delete Link - Part 2

+++ web-customer-tracker/WebContent/WEB-INF/view/list-customers.jsp
@@ -41,7 +41,7 @@
 						<td>
 							<a href="${updateLink}">Update</a>
 							|
-							<a href="${deleteLink}">Delete</a>
+							<a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this user?'))) return false;">Delete</a>
 						</td>
 					</tr>
 				</c:forEach>
